git-svn-id: svn://scit.us/dawg/trunk@157 7e5b8c5c-87e2-0310-a459-8d91125b37d5

+++ src/tree.cc
@@ -515,7 +515,7 @@ bool Tree::SetupEvolution(double pFreqs[], double pSubs[],
 	return true;
 }
 
-bool Tree::SetupRoot(const std::vector<std::string> &vSeqs, const std::vector<int> &vLens,
+bool Tree::SetupRoot(const std::vector<std::string> &vSeqs, const std::vector<unsigned long> &vLens,
 					   const std::vector<std::vector<double> > &vRates)
 {
 	m_vDNASeqs.clear();
